requirements.txt
# requirements.txt
      flask 
      sqlalchemy 
      flask_sqlalchemy 
      psycopg2-binary 
      flask_jwt_extended 
      flask_cors 
      requests 
      gunicorn
      Werkzeug
      python-dotenv
      python-dateutil
      celery
      pydantic
      sqlalchemy_utils
      redis
      rdbbeat
      Gevent
      Async_timeout
      celery-redbeat-2.3.3
      flask_mail-0.10.0
      blueprint-3.4.2
      Flask-Mail
      pycryptodome
